splits library from module
